Remove need for a ValidationWorker
